+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/security/SecurityConfig.java
@@ -1,8 +1,8 @@
 package org.eclipse.tractusx.puris.backend.common.security;
 
-import org.eclipse.tractusx.puris.backend.AuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
@@ -17,6 +17,11 @@ public class SecurityConfig {
         http.csrf((csrf) -> csrf.disable())
             .addFilterBefore(new AuthenticationFilter(), UsernamePasswordAuthenticationFilter.class)
             .authorizeRequests().anyRequest().authenticated();
+
+        http.csrf((csrf) -> csrf.disable())
+            .authorizeHttpRequests(
+                (authorizeObject) -> authorizeObject.anyRequest().authenticated()
+            ).httpBasic(Customizer.withDefaults());
 
         return http.build();
     }
